Edit(DEFLECT):ALLFunc

+++ ScoreManagement/Model/ScoreAnnoucement/ScoreAnnoucementResource.cs
@@ -18,9 +18,9 @@
         public string? firstname { get; set; }
         public string? lastname { get; set; }
         public string? seat_no { get; set; }
-        public int? accumulated_score { get; set; }
-        public int? midterm_score { get; set; }
-        public int? final_score { get; set; }
+        public decimal? accumulated_score { get; set; }
+        public decimal? midterm_score { get; set; }
+        public decimal? final_score { get; set; }
         public string? send_status_code { get; set; }
         public string? send_status_code_desc_th { get; set; }
         public string? send_status_code_desc_en { get; set; }
